+++ docs/CHANGELOG.md
@@ -6,6 +6,7 @@ Unreleased
 - Enh #68: Make URL locations clickable
 - Enh #69: Add `requirements.php`
 - Fix #71: After module disabling, Content for ExternalCalendarEntry are not deleted
+- Fix: Disabling the module in a container (Space or User)
 
 1.4.2 (January 19, 2023)
 -----------------------